requirements.txt
requests>=2.32.0
PyJWT>=2.8.0
tzdata>=2024.1
openai>=1.40.0
Pillow>=10.4.0
pytrends>=4.9.2
python-dotenv>=1.0.1

